MT-02-52

+++ controllers/itinerary.js
@@ -60,20 +60,28 @@ const controller = {
         let { id } = req.params;
 
         try {
-            let itinerary = await Itinerary.findOneAndUpdate({ _id: id }, req.body, { new: true });
-            if(itinerary){
-                res.status(200).json({
-                    success: true,
-                    message: 'Itinerary updated',
-                    data: itinerary,
-                });
-            }else{
-                res.status(404).json({
+            let oneItinerary = await Itinerary.findById(id)
+            if (oneItinerary.userId.equals(req.user.id)) {
+                let itinerary = await Itinerary.findOneAndUpdate({ _id: id }, req.body, { new: true });
+                if (itinerary) {
+                    res.status(200).json({
+                        success: true,
+                        message: 'Itinerary updated',
+                        data: itinerary,
+                    });
+                } else {
+                    res.status(404).json({
+                        success: false,
+                        message: 'Itinerary not found',
+                    });
+                }
+            } else {
+                res.status(401).json({
                     success: false,
-                    message: 'Itinerary not found',
+                    message: 'Unauthorized',
                 });
             }
-        }catch (error) {
+        } catch (error) {
             res.status(400).json({
                 success: false,
                 message: error.message,
@@ -83,22 +91,29 @@ const controller = {
 
     destroyOne: async (req, res) => {
         let { id } = req.params;
-
         try {
-            let itinerary = await Itinerary.findOneAndDelete({ _id: id });
-            if(itinerary){
-                res.status(200).json({
-                    success: true,
-                    message: 'Itinerary deleted',
-                    data: itinerary,
-                });
-            }else{
-                res.status(404).json({
+            let oneItinerary = await Itinerary.findById(id)
+            if (oneItinerary.userId.equals(req.user.id)) {
+                let itinerary = await Itinerary.findOneAndDelete({ _id: id });
+                if (itinerary) {
+                    res.status(200).json({
+                        success: true,
+                        message: 'Itinerary deleted',
+                        data: itinerary,
+                    });
+                } else {
+                    res.status(404).json({
+                        success: false,
+                        message: 'Itinerary not found',
+                    });
+                }
+            } else {
+                res.status(401).json({
                     success: false,
-                    message: 'Itinerary not found',
+                    message: 'Unauthorized',
                 });
             }
-    }catch (error) {
+        } catch (error) {
             res.status(400).json({
                 success: false,
                 message: error.message,
